refactor: extract interfaces to own file
